% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{Multiple Linear Regression.}
\arguments{
\item{formula}{Contains dependent and independent variables for linear regression}

\item{data}{A data.frame to conduct linear regression}
}
\description{
Multiple Linear Regression.
}

