% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\docType{class}
\name{ridgereg}
\alias{ridgereg}
\title{ridgereg}
\description{
Returns the result of the Linear Regression
}
\details{
Linear Regression
}
\section{Fields}{

\describe{
\item{\code{formula}}{formula. y ~ x_1 + x_2 + ... + x_n.}

\item{\code{data}}{data.frame.}
}}


